pyproject.toml
[project]
name = "deep-research-py"
version = "0.1.6"
description = "An AI-powered deep research assistant in python"
authors = [
    { name = "Esteban Puerta", email = "epuer94@gmail.com" }
]
dependencies = [
    "openai>=1.0.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.1",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.0.0",
    "firecrawl-py>=1.11.1",
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.0",
    "kitchenai-whisk",
]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "firecrawl-py"
]

[tool.hatch.build.targets.wheel]
packages = ["deep_research_py/"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["deep_research/"]
python_files = ["*_test.py"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true

[project.scripts]
deep-research = "deep_research_py.run:run"

[tool.uvicorn]
# Define the entry point for the uvx tool
entry-point = "deep_research_py.run:run"
